requirements.txt
streamlit==1.28.1
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1